Refactored Code and changed vertex.date type from int to float. Added another xml config and input examples (2D_Provenance)

+++ src/main/java/br/uff/ic/provviewer/Collapser.java
@@ -311,7 +311,7 @@ public class Collapser {
             Comparator comparator = new Comparator<Vertex>() {
                 @Override
                 public int compare(Vertex c1, Vertex c2) {
-                    return c1.getDate() - c2.getDate();
+                    return (int)(c1.getDate() - c2.getDate());
                 }
             };
             //Sort nodes by date
